Sync Pagination Changes
