--- at_secondary/at_secondary_server/lib/src/verb/handler/response/sync_stream_response_handler.dart
@@ -1,8 +1,0 @@
-import 'package:at_secondary/src/verb/handler/response/base_response_handler.dart';
-
-class SyncStreamResponseHandler extends BaseResponseHandler {
-  @override
-  String? getResponseMessage(String? verbResult, String prompt) {
-    return '';
-  }
-}
