--- at_secondary/at_secondary_server/lib/src/verb/handler/sync_stream_verb_handler.dart
@@ -1,72 +1,0 @@
-import 'dart:collection';
-import 'dart:convert';
-
-import 'package:at_commons/at_commons.dart';
-import 'package:at_persistence_secondary_server/at_persistence_secondary_server.dart';
-import 'package:at_secondary/src/server/at_secondary_impl.dart';
-import 'package:at_secondary/src/verb/handler/sync_verb_handler.dart';
-import 'package:at_secondary/src/verb/verb_enum.dart';
-import 'package:at_server_spec/at_server_spec.dart';
-import 'package:at_server_spec/at_verb_spec.dart';
-import 'package:at_server_spec/src/verb/verb.dart';
-
-class SyncStreamVerbHandler extends SyncVerbHandler {
-  static SyncStream syncStream = SyncStream();
-
-  SyncStreamVerbHandler(SecondaryKeyStore? keyStore) : super(keyStore);
-
-  @override
-  bool accept(String command) =>
-      command.startsWith(getName(VerbEnum.sync) + ':') && !command.startsWith('sync:from');
-
-  @override
-  Verb getVerb() {
-    return syncStream;
-  }
-
-  @override
-  Future<void> processVerb(
-      Response response,
-      HashMap<String, String?> verbParams,
-      InboundConnection? atConnection) async {
-    var commit_sequence = verbParams[AT_FROM_COMMIT_SEQUENCE]!;
-    var atCommitLog = await (AtCommitLogManagerImpl.getInstance()
-        .getCommitLog(AtSecondaryServerImpl.getInstance().currentAtSign));
-    var regex = verbParams[AT_REGEX];
-    var commit_changes =
-        atCommitLog?.getChanges(int.parse(commit_sequence), regex);
-    logger.finer(
-        'number of changes since commitId: $commit_sequence is ${commit_changes?.length}');
-    commit_changes?.removeWhere((entry) =>
-        entry.atKey!.startsWith('privatekey:') ||
-        entry.atKey!.startsWith('private:'));
-    if (regex != null && regex != 'null') {
-      logger.finer('regex for sync : $regex');
-      commit_changes
-          ?.removeWhere((entry) => !isRegexMatches(entry.atKey!, regex));
-    }
-    var distinctKeys = <String>{};
-    var syncResultList = [];
-    //sort log by commitId descending
-    commit_changes?.sort(
-        (entry1, entry2) => entry2.commitId!.compareTo(entry1.commitId!));
-    // for each latest key entry in commit log, get the value
-    if (commit_changes != null) {
-      await Future.forEach(commit_changes,
-          (dynamic entry) => processEntry(entry, distinctKeys, syncResultList));
-    }
-    logger.finer(
-        'number of changes after removing old entries: ${syncResultList.length}');
-    //sort the result by commitId ascending
-    syncResultList.sort(
-        (entry1, entry2) => entry1['commitId'].compareTo(entry2['commitId']));
-    if (syncResultList.isNotEmpty) {
-      syncResultList.forEach((element) {
-        logger.info('Syncing element on server ${element['atKey']}');
-        atConnection!
-            .write('${jsonEncode(element).length}~${jsonEncode(element)}\$');
-      });
-    }
-    return;
-  }
-}
