+++ at_server_spec/lib/verbs.dart
@@ -18,7 +18,6 @@ export 'package:at_server_spec/src/verb/stats.dart';
 export 'package:at_server_spec/src/verb/stream.dart';
 export 'package:at_server_spec/src/verb/sync.dart';
 export 'package:at_server_spec/src/verb/sync_from.dart';
-export 'package:at_server_spec/src/verb/sync_stream.dart';
 export 'package:at_server_spec/src/verb/update.dart';
 export 'package:at_server_spec/src/verb/update_meta.dart';
 export 'package:at_server_spec/src/verb/verb.dart';
